All works + cleaning
