+++ src/app/modules/analyst/components/report/malfunction-group-report/malfunction-group-report.component.html
@@ -1,35 +1,41 @@
 <mat-form-field>
-    <input matInput (keyup)="applyFilter($event.target.value)" placeholder="Filter">
-  </mat-form-field>
+  <input matInput (keyup)="applyFilter($event.target.value)" placeholder="Filter">
+</mat-form-field>
   
-  <div class="mat-elevation-z8">
-    <table mat-table [dataSource]="dataSource" matSort>
-  
+<div class="mat-elevation-z8">
+  <table mat-table [dataSource]="dataSource" multiTemplateDataRows  matSort>
 
-  
-      <ng-container *ngFor="let vehicleType of vehicleTypes; let i = index" [matColumnDef] = "vehicleType.name">
-        <th mat-header-cell *matHeaderCellDef mat-sort-header> {{vehicleType.name}} </th>
-        <td mat-cell *matCellDef="let row"> {{row.statistics[i]}} </td>
-      </ng-container>
-  
-      <ng-container matColumnDef="expandedDetail">
-        <td mat-cell *matCellDef="let element" [attr.colspan]="displayedColumns.length">
-          <div class="example-element-detail"
-               [@detailExpand]="element == expandedElement ? 'expanded' : 'collapsed'">
-               hoho
-          </div>
-        </td>
-      </ng-container>
+    <ng-container *ngFor="let columnName of displayedColumns; let i = index" [matColumnDef] = "columnName">
+      <th mat-header-cell *matHeaderCellDef mat-sort-header> {{columnName}} </th>
+      <td mat-cell *matCellDef="let row"> {{row[columnName]}} </td>
+    </ng-container>
 
-      <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
-      <tr mat-row *matRowDef="let row; columns: displayedColumns;"
+    <ng-container matColumnDef="expandedDetail">
+      <td mat-cell *matCellDef="let element" [attr.colspan]="displayedColumns.length">
+        <div class="example-element-detail"
+             [@detailExpand]="element == expandedElement ? 'expanded' : 'collapsed'">
+            <malfunction-subgroup-report [groupName] = "getCurrentGroup(element)"></malfunction-subgroup-report>
+
+        </div>
+      </td>
+    </ng-container>
+
+    <tr mat-header-row *matHeaderRowDef="displayedColumns"></tr>
+
+    <tr mat-row *matRowDef="let row; columns: displayedColumns;"
         class="example-element-row"
-        [class.example-expanded-row]="expandedElement === element"
-        (click)="expandedElement = expandedElement === element ? null : element">
-      </tr>
-    </table>
-  
-    <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]"></mat-paginator>
-  </div>
+        [class.example-expanded-row]="expandedElement === row"
+        (click)="expandedElement = expandedElement === row ? null : row">
+    </tr>
+
+    <tr mat-row *matRowDef="let row; columns: ['expandedDetail']" 
+        class="example-detail-row">
+    </tr>
+
+  </table>
+
+  <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]"></mat-paginator>
+
+</div>
