+++ src/app/modules/analyst/components/report/malfunction-group-report/malfunction-group-report.component.scss
@@ -1,13 +1,39 @@
-table {
-    width: 100%;
-  }
+mat-paginator {
+  background: #fffefee7;
+}
 
-.mat-form-field {
-  font-size: 14px;
+table {
+  width: 100%;
+  background: #fffefee7;
+}
+
+tr.example-detail-row {
+  height: 0;
+}
+
+tr.example-element-row:not(.example-expanded-row):hover {
+  background: rgba(255, 255, 255, 0.63)
+}
+
+tr.example-element-row:not(.example-expanded-row):active {
+  background: #efefef;
+}
+
+.example-element-row td {
+  border-bottom-width: 0;
+}
+
+.example-element-detail {
+  overflow: hidden;
   width: 100%;
 }
-  
-td, th {
-    width: 20%;
+
+.mat-column-expandedDetail {
+  padding: 0%;
+  margin: 0%;
 }
-  
+
+.mat-cell.cdk-column-expandedDetail.mat-column-expandedDetail.ng-tns-c6-0.ng-star-inserted {
+  padding: 0%;
+  margin: 0%;
+}
