+++ src/app/modules/analyst/components/report/malfunction-subgroup-report/malfunction-subgroup-report.component.html
@@ -34,7 +34,7 @@
 
   </table>
 
-  <mat-paginator [pageSizeOptions]="[1, 5, 10, 25, 100]"></mat-paginator>
+  <mat-paginator [pageSizeOptions]="[5, 10, 25, 100]"></mat-paginator>
 
 </div>
