+++ src/app/modules/analyst/components/report/malfunction-subgroup-report/malfunction-subgroup-report.component.scss
@@ -33,7 +33,7 @@ tr.example-element-row:not(.example-expanded-row):active {
   margin: 0%;
 }
 
-td.mat-cell.cdk-column-expandedDetail.mat-column-expandedDetail.ng-tns-c6-0.ng-star-inserted {
+.mat-cell.cdk-column-expandedDetail.mat-column-expandedDetail {
   padding: 0%;
   margin: 0%;
 }
