+++ src/app/modules/analyst/components/report/report.component.html
@@ -1,1 +1,1 @@
-<malfunction-subgroup-report groupName="Кузов">
+<malfunction-group-report></malfunction-group-report>
